Quiz study section refactored + file cleanup
A lot of changes have been made to files here! I refactored the Quiz code quite a bit and also cleaned up files (example we had duplicate images stored in two places and some files we no longer were using). I also renamed some of the routes and files for consistency across the project. Ideally and to ensure nothing else gets broken (if all looks good), we can just replace this branch with what's in main (not sure how that works best)? I still need to make the updates to Game section and work on sessions, but wanted to at least pull these updates now so we all can work on the most updated files!